Another CGI example.
Added another CGI example. This one calls a C program that gets the
system time and outputs a formatted-string representation of it. The
JavaScript then displays the formatted string onto the page. The next
step is to insert a timeout to call this C program repeatedly to see if
the time will update on the page.